flask_api/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
pandas==2.1.1
scikit-learn==1.3.1
numpy==1.26.0